+++ message/infrastructure/googlecloud/pubsub_test.go
@@ -12,7 +12,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/ThreeDotsLabs/watermill/message"
-	"github.com/ThreeDotsLabs/watermill/message/infrastructure/googlecloud"
 )
 
 // Run `docker-compose up` and set PUBSUB_EMULATOR_HOST=localhost:8085 for this to work
@@ -53,24 +52,6 @@ func TestPubsub(t *testing.T) {
 	t.Logf("Published a message with id %s on topic %s", id, topicName)
 }
 
-func TestNewPublisher(t *testing.T) {
-	ctx, _ := context.WithTimeout(context.Background(), 1*time.Second)
-
-	pub, err := googlecloud.NewPublisher(
-		ctx,
-		googlecloud.CreateTopicIfMissing(),
-	)
-	require.NoError(t, err)
-
-	topicName := fmt.Sprintf("test-topic-%d", rand.Int())
-
-	messages := []*message.Message{}
-	for i := 0; i < 5; i++ {
-		messages = append(messages,
-			message.NewMessage(uuid.NewV4().String(), message.Payload(msgText)),
-		)
-	}
-
-	err = pub.Publish(topicName, messages...)
-	require.NoError(t, err)
+func testMessage() *message.Message {
+	return message.NewMessage(uuid.NewV4().String(), message.Payload(msgText))
 }
